Added ability to open the app by double-clicking associated files.
If you publish the app and associate it with a filetype (e.g. png), double-clicking on that file type will launch the app and show the image.

To do:
Right now there's nothing to stop the app from trying to load file types it can't handle and it just silently fails.

The app should not open more than one instance, instead if the app is already open it should load any double-clicked images.

When installed the file associations should be set to the app, which a checkbox for each file type.

For any associated file types the icon should be set in Windows.

We still need to make the side panel files load asynchronously.

When opening an image the window should maximize.