Fullscreen now works better
I think there were issues in the last upload where fullscreen wouldn't always load correctly and the image could be slightly offscreen. Also if you had a 2nd monitor only the size of the primary monitor was taken into consideration and so fullscreen images could be the wrong size. Lastly the Windows Taskbar was always limiting the height.

All of the above was fixed.
Also started adding the ability to use the mouse wheel to zoom in and out of the image when in fullscreen mode. It works but it's not centering on the mouse yet, so that's the next thing to do.