Bug Fixes and Drag and Drop Implementation
Fixed a issue where the window was always on top of other windows after entering and exiting full screen mode.

Drag and drop now works as expected in both  windowed mode and full screen mode.

Note: Windows allows for multiple files to be dragged and dropped in one go, it might be a good idea to allow for this and have the thumbnails and next and previous be related to those files (and only those files).